refactor(docs-site): split policy-creator into tab shell + Playground
Pull the 14-section wizard body out of index.tsx and into a fresh
Playground component so the parent can host more than one mode.

index.tsx now owns:
  - the shared `policy` state (so both tabs see the same draft)
  - the active tab id (`quick-start` | `playground`), plus
  - a Quick Start `answers` blob (added in the next commit)
all three persist to localStorage so a refresh isn't destructive.

Playground itself is a near-verbatim move of the previous body. The
only behavioural addition is an optional `banner` slot the parent
fills with the "Loaded from Quick Start" handoff strip. The open-
section toggle gained a localStorage backing so the wizard remembers
which section you had open across reloads.

No section editor, validator, or emit pipeline changed.
